refactor : api 응답 유틸 메서드 활용 응답 로직 리팩토링
