+++ be/src/main/java/com/sharetravel/domain/user/controller/UserController.java
@@ -3,8 +3,8 @@ package com.sharetravel.domain.user.controller;
 import com.sharetravel.domain.user.dto.UserResponseDto;
 import com.sharetravel.domain.user.dto.UserUpdateRequestDto;
 import com.sharetravel.domain.user.service.UserService;
-import com.sharetravel.global.ApiResponseCode;
-import com.sharetravel.global.ApiResponseMessage;
+import com.sharetravel.global.api.ApiResponseCode;
+import com.sharetravel.global.api.ApiResponseMessage;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import static com.sharetravel.global.api.ApiUtil.getResponseEntity;
 
 @RequiredArgsConstructor
 @RestController
@@ -33,9 +35,7 @@ public class UserController {
 
     @ExceptionHandler(IllegalStateException.class)
     public ResponseEntity<ApiResponseMessage> handleInvalidTokenException() {
-        ApiResponseCode apiResponseCode = ApiResponseCode.USER_NOT_FOUND;
-        ApiResponseMessage apiResponseMessage = ApiResponseMessage.of(apiResponseCode.getCode(), apiResponseCode.getMessage());
-        return ResponseEntity.status(apiResponseCode.getHttpStatusCode()).body(apiResponseMessage);
+        return getResponseEntity(ApiResponseCode.USER_NOT_FOUND);
     }
 
     // TODO : 사용자 닉네임이 등록되지 않은 경우 클라이언트는 사용자가 닉네임을 등록하도록 하고 서버는 클라이언트로부터 사용자 닉네임을 받아 등록
